+++ AtomWeb.API/ATOM.Core/Services/IGatheringCenterService.cs
@@ -9,8 +9,8 @@ namespace ATOM.Core.Services
 {
     public interface IGatheringCenterService : IGenericService<GatheringCenter>
     {
-        Task<(BaseCenter, float distance)> NearGatheringCenter(float longitude, float latitude);
-        Task<List<BaseCenter>> ListNearGatheringCenters(float longitude, float latitude);
+        Task<(BaseCenter, float distance)> NearGatheringCenter(decimal longitude, decimal latitude);
+        Task<List<BaseCenter>> ListNearGatheringCenters(decimal longitude, decimal latitude);
 
     }
 }
