+++ AtomWeb.API/ATOM.Core/Services/IHelpDemandService.cs
@@ -11,7 +11,7 @@ namespace ATOM.Core.Services
     public interface IHelpDemandService : IGenericService<HelpDemand>
     {
         public Task AddHelpDemand(AddHelpDemandDto helpDemand);
-        public Task<(float AverageLatitude, float AverageLongitude)> AverageHelpLocation(int id);
+        public Task<(decimal AverageLatitude, decimal AverageLongitude)> AverageHelpLocation(int id);
 
     }
 }
